chore(arch-002): modularize api handlers and http utilities
