+++ apps/api/src/http/headers.ts
@@ -1,0 +1,6 @@
+import type { Headers } from "./types.js";
+
+export function getHeader(headers: Headers, key: string): string | null {
+  const value = headers[key.toLowerCase()] ?? headers[key];
+  return value ?? null;
+}
