+++ apps/api/src/http/types.ts
@@ -1,0 +1,6 @@
+export type Headers = Record<string, string | undefined>;
+
+export interface ApiResponse {
+  status: number;
+  body: unknown;
+}
